Django Rest Api/requirements.txt
asgiref==3.7.2
Django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==21.0.0
more-itertools==10.1.0
numpy==1.26.2
pandas==2.1.3
portalocker==2.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
runscript==0.3.5
setproctitle==1.3.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
